+++ tests/Umbraco.Tests.Common/Builders/ContentTypeBuilder.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Umbraco.
 // See LICENSE for more details.
 
-using System.Collections.Generic;
-using System.Linq;
 using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Models.ContentEditing;
@@ -125,6 +123,7 @@ public class ContentTypeBuilder
         contentType.Trashed = GetTrashed();
         contentType.ListView = GetListView();
         contentType.IsElement = _isElement ?? false;
+        contentType.AllowedAsRoot = GetAllowedAtRoot();
         contentType.HistoryCleanup = new HistoryCleanup();
 
         contentType.Variations = contentVariation;
